# docs/openapi.yaml
openapi: 3.0.1
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /rule:
    get:
      tags:
      - rule-controller
      operationId: getAll
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
    post:
      tags:
      - rule-controller
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DynamicRuleDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DynamicRuleDto"
...
  /rule/{productId}:
    delete:
      tags:
      - rule-controller
      operationId: delete
      parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
components:
  schemas:
    DynamicRuleDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        productId:
          type: string
        productName:
          type: string
        productText:
          type: string
        rule:
          type: array
          items:
            type: object
            additionalProperties:
              type: object
...